net/mlx5e: Remove redundant WARN when we cannot find neigh entry
It is possible for neigh entry not to exist if it was cleaned already.
When we bring down an interface the neigh gets deleted but it could be
that our listener for neigh event to clear the encap valid bit didn't
start yet and the neigh update last used work is started first.
In this scenario the encap entry has valid bit set but the neigh entry
doesn't exist.

Issue: 1324532
Change-Id: Idea557f8fcb10287f3db544c6a5fdddda844d53f
Signed-off-by: Roi Dayan <roid@mellanox.com>
Reviewed-by: Paul Blakey <paulb@mellanox.com>

[juergh: IBM patchset mlx-4.15.0-24.26-patches-20180921 (SF #00199436).]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>