KVM: PPC: Book3S HV: Radix page fault handler optimizations
BugLink: https://bugs.launchpad.net/bugs/1800649

This improves the handling of transparent huge pages in the radix
hypervisor page fault handler.  Previously, if a small page is faulted
in to a 2MB region of guest physical space, that means that there is
a page table pointer at the PMD level, which could never be replaced
by a leaf (2MB) PMD entry.  This adds the code to clear the PMD,
invlidate the page walk cache and free the page table page in this
situation, so that the leaf PMD entry can be created.

This also adds code to check whether a PMD or PTE being inserted is
the same as is already there (because of a race with another CPU that
faulted on the same page) and if so, we don't replace the existing
entry, meaning that we don't invalidate the PTE or PMD and do a TLB
invalidation.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
(cherry picked from commit c4c8a7643e74ebd7f2cfa80807562f16bb58c1d9)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>
[juergh: IBM patchset nvlink2-based-on-ibmgt.tar.gz (SF #00213341).]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>