UBUNTU: [Packaging] config-check: allow overlay annotations files
BugLink: http://bugs.launchpad.net/bugs/1752072

Update the config-check script to parse additional annotations files
that can overlay the original annotations on a config basis. The overlay
files are read from "$DEBIAN/config/*.annotations".

Since the overlay works on a config basis, any mention to a config in
the overlay file will cause the entry for that config on the original
file to be ignore.

For instance, the following line in the overlay file:

CONFIG_X note<some note>

Will cause any line related to CONFIG_X in the original annotations file
to be ignored.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>