KVM: PPC: return EINVAL for HPT guest with indep_threads_mode=Y
In order to run HPT guests on radix hosts KVM must be loaded with
indep_threads_mode=N argument. The reason is that POWER9 expects all
SMT threads of a given core to run the same MMU mode (radix or hash),
which prevents the usage of different MMU contexts independently.

To address the problem the system must run in single threaded mode and
kvm_hv module loaded with indep_threads_mode=N argument. Thus, KVM is
able to know when to manage individual threads or the whole set of
threads.

The problem is that kvm_hv returns a generic error code (EBUSY) if
users try to run HPT guests in radix host with indep_threads_mode=Y
(default). Thus, the client isn't able to display a meaningful error
message to the user.

This patch simply return EINVAL, that indicates a wrong argument
for indep_threads_mode.

Signed-off-by: Jose Ricardo Ziviani <joserz@linux.vnet.ibm.com>