net/mlx5: Add source e-switch owner
The source e-switch owner allows a vport on one e-switch port be associated
with a rule defined on the second port e-switch.

The role of the source eswitch owner valid bit in the flow group is to
allow the firmware fail driver attempts to wild card the source eswitch
match field. If this bit is not set, the firmware ignores the source
eswitch owner field totally.

Issue: 1159068
Change-Id: I368ee3ea45b6c527a3ba5cd0fa6973b210697c86
Signed-off-by: Shahar Klein <shahark@mellanox.com>
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>
Reviewed-by: Roi Dayan <roid@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>

[juergh: IBM patchset mlx-4.15.0-24.26-patches-20180921 (SF #00199436).]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>