vfio/pci: Quiet broken INTx whining when INTx is unsupported by device
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

The intent of commit 2170dd04316e ("vfio-pci: Mask INTx if a device is
not capabable of enabling it") was to disallow the user from seeing
that the device supports INTx if the platform is incapable of enabling
it.  The detection of this case however incorrectly includes devices
which natively do not support INTx, such as SR-IOV VFs.  We don't need
the nointx handling code for such devices, nor do we want to issue a
broken INTx masking warning.  Check whether the device supports INTx
before enabling nointx and remove the free pass for non-i40e Intel
devices.

Reported-by: Arjun Vynipadath <arjun@chelsio.com>
Fixes: 2170dd04316e ("vfio-pci: Mask INTx if a device is not capabable of enabling it")
Acked-by: Casey Leedom <leedom@chelsio.com>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit bcd1face5ce84300b8d3d2a5bee26452ed0d33e2)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>