net/mlx5e: Properly order min inline mode setup while parsing TC matches
Set the initial value to none instead of L2, and set to L2
where the previous initial value was assumed. Make sure to
parse L2 matches before L3 matches and L3 before L4.

This is a pre-step to get the match level for more purposes
other than the validating the needed vs. actual inline level.

Issue: 1338214
Change-Id: Iaac8ca98654099fae74eb2d7d2f42968bba37c37
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Reviewed-by: Roi Dayan <roid@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>

[juergh: IBM patchset mlx-4.15.0-24.26-patches-20180921 (SF #00199436).]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>