powerpc/powernv/npu: Add a debugfs setting to change ATSD threshold
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

The threshold at which it becomes more efficient to coalesce a range
of ATSDs into a single per-PID ATSD is currently not well understood
due to a lack of real-world work loads. This patch adds a debugfs
parameter allowing the threshold to be altered at runtime in order to
aid future development and refinement of the value.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Acked-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(cherry picked from commit 99c3ce33a00bc40cb218af770ef00c82c8044c36)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>