vfio_pci: Allow mapping extra regions
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

So far we only allowed mapping of MMIO BARs to the userspace. However
there there are GPUs with on-board coherent RAM accessible via side
channels which we also want to map to the userspace. The first client
for this is NVIDIA V100 GPU with NVLink2 direct links to a POWER9
NPU-enabled CPU; such GPUs have 16GB RAM which is coherently mapped
to the system address space, we are going to export these as an extra
PCI region.

We already support extra PCI regions and this adds support for mapping
them to the userspace.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
---
Changes:
v2:
* reverted one of mistakenly removed error checks
(cherry picked from commit 4a080e1e2c47f21cdb8b974764cf422369ecdd46)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>