vfio/pci: Make IGD support a configurable option
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

Allow the code which provides extensions to support direct assignment
of Intel IGD (GVT-d) to be compiled out of the kernel if desired.  The
config option for this was previously automatically enabled on X86,
therefore the default remains Y.  This simply provides the option to
disable it even for X86.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit 08ca1b52f69b4dfa8703d54e26e2c6e11aa453eb)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>