powerpc/powernv: Move npu struct from pnv_phb to pci_controller
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

The powernv PCI code stores NPU data in the pnv_phb struct. The latter
is referenced by pci_controller::private_data. We are going to have NPU2
support in the pseries platform as well but it does not store any
private_data in in the pci_controller struct; and even if it did,
it would be a different data structure.

This makes npu a pointer and stores it one level higher in
the pci_controller struct.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v4:
* changed subj from "powerpc/powernv: Detach npu struct from pnv_phb"
* got rid of global list of npus - store them now in pci_controller
* got rid of npdev_to_npu() helper
(cherry picked from commit de10d25f137842fad9c18f56e8e1f67ced8547ee)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>