vfio_pci: Allow regions to add own capabilities
BugLink: https://bugs.launchpad.net/bugs/1800649

VFIO regions already support a region capability; however the subdriver
might have to report to the userspace more than that, for example,
the base address of the region in the real hardware is needed by
the NVIDIA V100 GPU (connected to a CPU via NVLink) in order to dispatch
untranslated DMA trafic to the correct NVLink.

This adds an add_capability() hook to vfio_pci_regops.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>