powerpc/pseries/npu: Enable platform support
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

We already changed NPU API for GPUs to not to call OPAL and the remaining
bit is initializing NPU structures.

This searches for POWER9 NVLinks attached to any device on a PHB and
initializes an NPU structure if any found.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v4:
* dropped "IBM,npu-vphb" compatible type on PHB and use the type of NVLink
(cherry picked from commit a815569093d4579ccbbd9a390f17bdcf70216ac7)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>