FIXME: powerpc/powernv: Detach npu struct from pnv_phb
BugLink: https://bugs.launchpad.net/bugs/1800649

The powernv PCI code stores NPU data in the pnv_phb struct reference to
which is stored in pci_controller::private_data.

We are going to have NPU2 support in pseries platform as well but it does
not store any private_data in in the pci_controller struct.

This adds a global list of NPUs so each platform can register and use
these in the same fashion.

This moves pnv_npu2_init() declaration out of the powernv folder so
later pseries could include it; this also renames it.

This moves LPAR mapping to a separate helper as later we are going to
call pnv_npu2_init() from the pseries platform and LPAR is going to get
mapped differently.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>