vfio_pci: Add NVIDIA GV100GL [Tesla V100 SXM2] [10de:1db1] subdriver
BugLink: https://bugs.launchpad.net/bugs/1800649

Some POWER9 chips come with special NVLink2 links which provide
cacheable memory access to the RAM physically located on NVIDIA GPU.
This memory is presented to a host via the device tree but remains
offline until the NVIDIA driver onlines it.

This exports this RAM to the userspace as a new region so
the NVIDIA driver in the guest can train these links and online GPU RAM.

This exports ATSD register to do whatever the driver wants to do with it
on a bare metal system.

This does want KVM as the nest MMU needs to be told about LPID
(logical partition ID, a guest ID) for ATS and there is no known NVIDIA
driver version supporting no-ATS or no-coherent-RAM configurations.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>