KVM: PPC: Book3S HV: Add a debugfs file to dump radix mappings
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

This adds a file called 'radix' in the debugfs directory for the
guest, which when read gives all of the valid leaf PTEs in the
partition-scoped radix tree for a radix guest, in human-readable
format.  It is analogous to the existing 'htab' file which dumps
the HPT entries for a HPT guest.

Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(cherry picked from commit 9a94d3ee2d159927c0f8e5078228eadbce8dda43)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>