KVM: PPC: Use ccr field in pt_regs struct embedded in vcpu struct
BugLink: https://canonical.my.salesforce.com/5003z00001xnYEC (KVM: Back port POWER9 nested virtualization patches (SF: #00228446))

When the 'regs' field was added to struct kvm_vcpu_arch, the code
was changed to use several of the fields inside regs (e.g., gpr, lr,
etc.) but not the ccr field, because the ccr field in struct pt_regs
is 64 bits on 64-bit platforms, but the cr field in kvm_vcpu_arch is
only 32 bits.  This changes the code to use the regs.ccr field
instead of cr, and changes the assembly code on 64-bit platforms to
use 64-bit loads and stores instead of 32-bit ones.

Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Leonardo Garcia <lagarcia@br.ibm.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>