KVM: PPC: Book3S HV: Nested guest entry via hypercall
BugLink: https://canonical.my.salesforce.com/5003z00001xnYEC (KVM: Back port POWER9 nested virtualization patches (SF: #00228446))

This adds a new hypercall, H_ENTER_NESTED, which is used by a nested
hypervisor to enter one of its nested guests.  The hypercall supplies
register values in two structs.  Those values are copied by the level 0
(L0) hypervisor (the one which is running in hypervisor mode) into the
vcpu struct of the L1 guest, and then the guest is run until an
interrupt or error occurs which needs to be reported to L1 via the
hypercall return value.

Currently this assumes that the L0 and L1 hypervisors are the same
endianness, and the structs passed as arguments are in native
endianness.  If they are of different endianness, the version number
check will fail and the hcall will be rejected.

Nested hypervisors do not support indep_threads_mode=N, so this adds
code to print a warning message if the administrator has set
indep_threads_mode=N, and treat it as Y.

Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(backported from commit 360cae313702cdd0b90f82c261a8302fecef030a)
Signed-off-by: Fabiano Rosas <farosas@linux.ibm.com>
Signed-off-by: Leonardo Garcia <lagarcia@br.ibm.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>