powerpc/iommu_api: Move IOMMU groups setup to a single place
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1800649

Registering new IOMMU groups and adding devices to them are separated in
code and the latter is dug in the DMA setup code which it does not
really belong to.

This moved IOMMU groups setup to a separate helper which registers a group
and adds devices as before. This does not make a difference as IOMMU
groups are not used anyway; the only dependency here is that
iommu_add_device() requires a valid pointer to an iommu_table
(set by set_iommu_table_base()).

To keep the old behaviour, this does not add new IOMMU groups for PEs
with no DMA weigth and also skips NVLink bridges which do not have
pci_controller_ops::setup_bridge (the normal way of adding PEs).

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
(cherry picked from commit f1b3b269cb48c10bc793c5c7ee6330ee66895f0a)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>