net/mlx5: E-Switch, Reload IB interface when switching devlink modes
Up until this point it wasn't possible to activate IB representors
when switching to switchdev mode, remove this limitation.

We trigger reload of the PF IB interface in order to make sure that
already allocated resources are invalid and new resources will be opened
correctly with all the limitations of switchdev mode applied (only raw
packet capabilities, without RoCE).

Signed-off-by: root <root@clx-ibmc-02.mtl.labs.mlnx>

[juergh: IBM patchset mlx-4.15.0-24.26-patches-20181108 (SF #00204763).]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>