powerpc/powernv/npu: Move single TVE handling to NPU PE
Normal PCI PEs have 2 TVEs, one per a DMA window; however NPU PE has only
one which points to one of two tables of the corresponding PCI PE.

So whenever a new DMA window is programmed to PEs, the NPU PE needs to
release old table in order to use the new one.

Commit d41ce7b1bcc3e ("powerpc/powernv/npu: Do not try invalidating 32bit
table when 64bit table is enabled") did just that but in pci-ioda.c
while it actually belongs to npu-dma.c.

This moves the single TVE handling to npu-dma.c. This does not implement
restoring as it is going to be done in more generic way in following
patches.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
(cherry picked from commit 67cbfd38731c355fa96358e302736b05028bba28)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.ibm.com>